+++ ZImageViewer/ImageViewer.cpp
@@ -1183,7 +1183,7 @@ void ImageViewer::LimitIndex()
 
 void ImageViewer::UpdateControlPanel()
 {
-    int64_t nControlPanelSide = gM * 2.5;
+    int64_t nControlPanelSide = (int64_t)(gM * 2.5);
     limit<int64_t>(nControlPanelSide, 40, 88);
